requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.5
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
Mako==1.3.7
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-jose==3.3.0
python-multipart==0.0.19
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
sqlalchemy-views==0.3.2
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
WTForms==3.1.2
